FON-105 Create add method and fix bugs

+++ src/main/java/com/fontana/backend/bootstrap/BootstrapInMemoryData.java
@@ -106,7 +106,7 @@ public class BootstrapInMemoryData implements CommandLineRunner {
                 .username(userRepository.findAll().get(0).getUsername())
                 .sessionId(sessionRepository.findAll().get(0).getId())
                 .deviceType(DeviceType.PUMP)
-                .deviceValue((short) 48)
+                .deviceValue(48)
                 .build();
 
         logRepository.save(first);
